agents/citation_manager/requirements.txt
fastapi
uvicorn[standard]
httpx
requests
jsonschema
citeproc-py
lxml
pdfplumber
python-multipart
pymupdf
